v1.2.1 重写了AES加密解密、修复FreebusTransport打印多余日志问题、修复YzamiMaster通讯异常导致的数值越界问题。新增一些扩展方法

+++ Palink.Tools.Core/Freebus/Interface/IFreebusTransport.cs
@@ -7,9 +7,9 @@ namespace Palink.Tools.Freebus.Interface;
 
 public interface IFreebusTransport : IDisposable
 {
-    int Retries { get; set; }
+    ushort Retries { get; set; }
 
-    int WaitToRetryMilliseconds { get; set; }
+    ushort WaitToRetryMilliseconds { get; set; }
 
     int ReadTimeout { get; set; }
